refactor: [#1491] put implementation and type in the same module
