+++ packages/torrent-repository/src/entry/mod.rs
@@ -1,3 +1,2 @@
 pub mod peer_list;
-pub mod single;
 pub mod torrent;
